% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process USDA JSON data}
\usage{
process_data(data)
}
\arguments{
\item{data}{A data file in json format}
}
\description{
This function takes json data from the USDA food nutrient API and returns essential information
}
\examples{
process_data()
}
\keyword{json}
