% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_usda.R
\name{query_usda}
\alias{query_usda}
\title{Queries the USDA database}
\usage{
query_usda(codes, api.key, format = "json")
}
\arguments{
\item{codes}{A vector of USDA codes}

\item{api.key}{A user's API key}

\item{format}{The format of the data to pull. Defaults to json}
}
\description{
Runs the query to the USDA database
}
\examples{
query_usda()
}
\keyword{json}
