% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usda_api.R
\name{usda_api}
\alias{usda_api}
\title{Sets up the call to the USDA API}
\usage{
usda_api(path)
}
\arguments{
\item{path}{A defined path to append to the set API}
}
\description{
Defines and passes the string to pull from USDA
}
\examples{
process_data()
}
\keyword{json}
